# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/loan/{id}": {
            "get": {
                "description": "Show an Loan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan"
                ],
                "summary": "Show Loan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Show Loan Request",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ShowLoanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_loan.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_loan.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an Loan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan"
                ],
                "summary": "Delete an Loan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Delete Loan Param",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DeleteLoanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_loan.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_loan.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/transaction": {
            "put": {
                "description": "Update a Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Update a Transaction",
                "parameters": [
                    {
                        "description": "Update Transaction Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_transaction.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_transaction.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_transaction.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Create a Transaction",
                "parameters": [
                    {
                        "description": "Create Transaction Params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_transaction.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_transaction.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transaction/{id}": {
            "get": {
                "description": "Show an transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Show a transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Show Transaction Request",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ShowTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_transaction.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_transaction.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Delete a transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Delete Transaction Param",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DeleteTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_transaction.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_transaction.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/{id}": {
            "get": {
                "description": "Show an user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Show user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Show User Request",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ShowUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_user.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_user.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Delete an user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Delete User Param",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DeleteUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_user.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_user.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.Loan": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "bank_account_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "interest_rate": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "db.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "bank_account_id": {
                    "type": "integer"
                },
                "credit_debit": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "db.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_isaiaspereira307_gowallet_handlers_loan.ErrorResponse": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_isaiaspereira307_gowallet_handlers_transaction.ErrorResponse": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_isaiaspereira307_gowallet_handlers_user.ErrorResponse": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.CreateTransactionRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "bank_account_id": {
                    "type": "integer"
                },
                "credit_debit": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                }
            }
        },
        "handlers.CreateTransactionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Transaction"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.DeleteLoanResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Loan"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.DeleteTransactionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Transaction"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.DeleteUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.User"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ShowLoanResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Loan"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.ShowTransactionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Transaction"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.ShowUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.User"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.UpdateTransactionRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "amount": {
                    "type": "string"
                },
                "bank_account_id": {
                    "type": "integer"
                },
                "credit_debit": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "handlers.UpdateTransactionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Transaction"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}